+++ modules-template/prov-template-compiler/src/main/java/module-info.java
@@ -19,6 +19,7 @@ module prov.template.compiler {
     requires prov.template;
 
 
+
     exports org.openprovenance.prov.template.compiler;
     exports org.openprovenance.prov.template.compiler.configuration;
     exports org.openprovenance.prov.template.compiler.util;